AgenticWorkers/QueryAnalyst/requirements.txt
python-dotenv
psycopg2-binary
sentence-transformers
Pillow
requests
google-generativeai
groq
langgraph
reportlab
azure-storage-queue
azure-storage-blob
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
crewai
tavily-python